Resolve some warnings

+++ Erathosthenes/known_primes.h
@@ -9,7 +9,7 @@ std::vector<uint32_t> get_known_primes(size_t limit)
 	std::cout << "Calculating base primes" << std::endl;
 
 	auto knownPrimes = std::vector<uint32_t>();
-	knownPrimes.reserve(limit / log(limit));
+	knownPrimes.reserve(limit / static_cast<size_t>(log(limit)));
 	knownPrimes.push_back(2u);
 	for (auto i = 3u; i < limit; i += 2u)
 	{
@@ -33,7 +33,7 @@ std::vector<uint32_t> get_known_primes(size_t limit)
 
 void update_known_primes(std::vector<uint32_t>& known_primes, const Array<bool>& result, size_t limit)
 {
-	known_primes.reserve(limit / log(limit));
+	known_primes.reserve(limit / static_cast<size_t>(log(limit)));
 	for (auto i = known_primes.back() + 2u; i < limit; i += 2u)
 	{
 		if (result[i])
